Move alignment types to own type file.
