+++ src/tools/components/__tests__/AlignmentView.spec.tsx
@@ -2,7 +2,8 @@ import { screen, fireEvent } from '@testing-library/react';
 
 import customRender from '../../../shared/__test-helpers__/customRender';
 
-import AlignmentView, { MSAInput, Tool, View } from '../AlignmentView';
+import AlignmentView, { Tool, View } from '../AlignmentView';
+import { MSAInput } from '../../types/alignment';
 
 import mockData from '../__mocks__/msaMocks.json';
